ppc: Improve PCR bit selection in ppc_set_compat()
RH-Author: Thomas Huth <thuth@redhat.com>
Message-id: <1466022918-18958-5-git-send-email-thuth@redhat.com>
Patchwork-id: 70629
O-Subject: [RHEL-7.3 qemu-kvm-rhev PATCH 4/5] ppc: Improve PCR bit selection in ppc_set_compat()
Bugzilla: 1341492
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>

When using an olderr PowerISA level, all the upper compatibility
bits have to be enabled, too. For example when we want to run
something in PowerISA 2.05 compatibility mode on POWER8, the bit
for 2.06 has to be set beside the bit for 2.05.
Additionally, to make sure that we do not set bits that are not
supported by the host, we apply a mask with the known-to-be-good
bits here, too.

Signed-off-by: Thomas Huth <thuth@redhat.com>
[dwg: Added some #ifs to fix compile on 32-bit targets]
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>

(cherry picked from commit eac4fba965136f61cc239a450bec12adcef6b449)
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>