Revert "e1000: use alias for default model"
RH-Author: Dr. David Alan Gilbert <dgilbert@redhat.com>
Message-id: <1467803295-18634-3-git-send-email-dgilbert@redhat.com>
Patchwork-id: 70969
O-Subject: [RHEL-7.3 qemu-kvm-rhev PATCH 2/3] Revert "e1000: use alias for default model"
Bugzilla: 1353070
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Xiao Wang <jasowang@redhat.com>

From: "Dr. David Alan Gilbert" <dgilbert@redhat.com>

This reverts commit 8304402033e8dbe8e379017d51ed1dd8344f1dce.

In 7.x we have a compat entry for e1000 to use the rhel6-e1000.rom,
but we never added the same entry for e1000-82540em (which rhel6 doesn't
have).
Converting to an alias, which this patch did, means that the compat
entry gets used for both e1000 and e1000-82540em and so you end up
with a different ROM when going 7.2->7.3 with the rhel6.6.0 machine type
and the e1000-82540em device.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>