crypto: fix handling of iv generator hash defaults
RH-Author: Daniel P. Berrange <berrange@redhat.com>
Message-id: <1472312195-23794-5-git-send-email-berrange@redhat.com>
Patchwork-id: 72101
O-Subject: [PATCH qemu-kvm-rhev 7.3 4/5] crypto: fix handling of iv generator hash defaults
Bugzilla: 1301019
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

When opening an existing LUKS volume, if the iv generator is
essiv, then the iv hash algorithm is mandatory to provide. We
must report an error if it is omitted in the cipher mode spec,
not silently default to hash 0 (md5).  If the iv generator is
not essiv, then we explicitly ignore any iv hash algorithm,
rather than report an error, for compatibility with dm-crypt.

When creating a new LUKS volume, if the iv generator is essiv
and no iv hsah algorithm is provided, we should default to
using the sha256 hash.

Reported-by: Paolo Bonzini <pbonzini@redhat.com>
Reviewed-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
(cherry picked from commit 8b7cdba386d55ecee2caa26973c1d6c31822e801)
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>