crypto: ensure XTS is only used with ciphers with 16 byte blocks
RH-Author: Daniel P. Berrange <berrange@redhat.com>
Message-id: <1472312195-23794-6-git-send-email-berrange@redhat.com>
Patchwork-id: 72102
O-Subject: [PATCH qemu-kvm-rhev 7.3 5/5] crypto: ensure XTS is only used with ciphers with 16 byte blocks
Bugzilla: 1301019
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

The XTS cipher mode needs to be used with a cipher which has
a block size of 16 bytes. If a mis-matching block size is used,
the code will either corrupt memory beyond the IV array, or
not fully encrypt/decrypt the IV.

This fixes a memory corruption crash when attempting to use
cast5-128 with xts, since the former has an 8 byte block size.

A test case is added to ensure the cipher creation fails with
such an invalid combination.

Reviewed-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
Upstream:
  - pending for 2.8
  - http://lists.nongnu.org/archive/html/qemu-devel/2016-08/msg03943.html

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>