io: remove mistaken call to object_ref on QTask
RH-Author: Daniel P. Berrange <berrange@redhat.com>
Message-id: <1473850467-30763-1-git-send-email-berrange@redhat.com>
Patchwork-id: 72315
O-Subject: [RHEV-7.3 qemu-kvm-rhev PATCH] io: remove mistaken call to object_ref on QTask
Bugzilla: 1375677
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Eric Blake <eblake@redhat.com>

The QTask struct is just a standalone struct, not a QOM Object,
so calling object_ref() on it is not appropriate. This results
in mangling the 'destroy' field in the QTask struct, causing
the later call to qtask_free() to try to call the function
at address 0x1, with predictably segfault happy results.

There is in fact no need for ref counting with QTask, as the
call to qtask_abort() or qtask_complete() will automatically
free associated memory.

This fixes the crash shown in

  https://bugs.launchpad.net/qemu/+bug/1589923

Reviewed-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
(cherry picked from commit bc35d51077b33e68a0ab10a057f352747214223f)
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>