vus: implement qemu heartbeat on vus socket
Summary:
Implements a heartbeat on the vhost-user-scsi unix socket by querying
Frodo's features every 5 seconds. If the socket is closed, then the
read/write will fail with -1 which prompts Qemu to abort().

A further improvement can be achieved by setting an alarm() before
sending the heartbeat (and clearing it on receive). That will also catch
cases where Frodo is not responding, but still alive.

Signed-off-by: Felipe Franciosi <felipe@nutanix.com>
Signed-off-by: Peter Turschmid <peter.turschm@nutanix.com>
Test Plan: dev
Reviewers: cui
Reviewed By: cui
CC: ahv
Differential Revision: http://codereview.nutanix.net/D14128