Remove dirty_rate_high_cnt
Summary:
The dirty_rate_high_cnt counter is used by Qemu to retry migration
iterations a few times before increasing the throttling level. Due
to a few bugs:
1) The first iteration is retried four times;
2) Every other iteration is retried three times.

The original commit for this code (070afca2) suggests it should only
attempt twice. Given we are going to throttle the guest, even attempting
twice is a waste of network bandwidth and unecessary torture to the
workload. This commit just drops this logic altogether.
Test Plan: dev
Reviewers: malcolm.crossley, igrobman, david.vrabel, fabien.hermenier, prerna.saxena, aramesh
CC: ahv
Differential Revision: http://codereview.nutanix.net/D14312