Throttle based on dirty/xfer rates
Summary:
Original Qemu throttles by fixed increments while the dirty rate is more
than half of the xfer rate. This changes the approach to throttle
straight to the amount needed so that the dirty/xfer rate criteria is
met. It also fixes a bug that incorrectly initialised bytes_xfer_prev.

Signed-off-by: Felipe Franciosi <felipe@nutanix.com>
Signed-off-by: David Vrabel <david.vrabel@nutanix.com>
Signed-off-by: Malcolm Crossley <malcolm@nutanix.com>
Test Plan: dev
Reviewers: malcolm.crossley, igrobman, david.vrabel, fabien.hermenier, prerna.saxena, aramesh, cui
CC: ahv
Reviewed-by: igrobman, malcolm.crossley
Differential Revision: http://codereview.nutanix.net/D14373