Don't iterate twice the first time around
Summary:
Qemu only calculates the rate counters (eg. dirty_pages_rate) _after_ it
has a first go at the auto convergence logic. This is obviously broken,
since it cannot do any auto convergence calculation unless it has such
values. This commit moves the rate counters update to before the auto
convergence logic in order to fix this.

This also gets rid of the unecessary check on a zero counter: the
counter will only be zero if no pages were dirtied, in which case the
conditional result will be the same.
Test Plan: dev
Reviewers: malcolm.crossley, igrobman, david.vrabel, fabien.hermenier, prerna.saxena, aramesh, cui
CC: ahv
Differential Revision: http://codereview.nutanix.net/D14323