Only mark throttle as completed after sleep
Summary:
Currently qemu will set cpu->throttle_thread_scheduled to 0 before
taking a vcpu to sleep as part of throttling. This means that if the
cpu_throttle_timer_tick timer ticks again before the sleep completes,
the throttle thread will run again immediately after it finishes,
without actually allowing the vcpu to run normally in between.

The result is several consecutive sleeps applied to a vcpu, potentially
not running the guest for several seconds. This symptom was observed
when throttling the guest at 99%, which makes the scenario above more
likely to happen.
Test Plan: dev
Reviewers: malcolm.crossley, igrobman, david.vrabel, fabien.hermenier, prerna.saxena, aramesh, cui
CC: ahv
Reviewed-by: igrobman, malcolm.crossley
Differential Revision: http://codereview.nutanix.net/D14374