target-ppc: Handle additional bits in H_GET_CPU_CHARACTERISTICS
The hcall H_GET_CPU_CHARACTERISTICS is used by the guest to determine
information about the required behaviours and available characteristics
of the cpu.

Bits may be added to this hcall, the information for which is queried
from the hypervisor using the KVM_PPC_GET_CPU_CHAR ioctl. In order to
handle the case where additional bits may be added to this hcall, and
thus to the ioctl, relax the parsing in qemu so that there is minimal
modification of the data before it is passed to the guest.

Signed-off-by: Suraj Jitindar Singh <sjitindarsingh@gmail.com>
Signed-off-by: Ricardo M. Matinata <rmm@br.ibm.com>