target-i386: cpu: Add downstream-only STIBP CPUID flag
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <20180926185059.20691-1-ehabkost@redhat.com>
Patchwork-id: 82301
O-Subject: [RHEL-7.6 qemu-kvm-rhev PATCH] target-i386: cpu: Add downstream-only STIBP CPUID flag
Bugzilla: 1638077
RH-Acked-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
RH-Acked-by: Kashyap Chamarthy <kchamart@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>

From: Paolo Bonzini <pbonzini@redhat.com>

We accidentally dropped the downstream-only STIBP CPUID flag
during the 2.12.0 rebase.

STIBP is a CPUID flag that was considered for the Spectre
(CVE-2017-5715) mitigations, but in the end it was not necessary:
spec-ctrl/IBRS was deemed enough.  The kernel KVM STIBP CPUID
code was never merged upstream, but it's present on RHEL-7.

This means we may have existing VMs created on RHEL-7.5 hosts
with the STIBP flag enabled, and we need to support
live-migration of those VMs to RHEL-7.6.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>