scsi-generic: avoid out-of-bounds access to VPD page list
RH-Author: Paolo Bonzini <pbonzini@redhat.com>
Message-id: <20181107180007.22954-8-pbonzini@redhat.com>
Patchwork-id: 82939
O-Subject: [RHEL7.6.z qemu-kvm-rhev PATCH 7/9] scsi-generic: avoid out-of-bounds access to VPD page list
Bugzilla: 1640927
RH-Acked-by: Max Reitz <mreitz@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>

A device can report an excessive number of VPD pages when asked for a
list; this can cause an out-of-bounds access to buf in
scsi_generic_set_vpd_bl_emulation.  It should not happen, but
it is technically not incorrect so handle it: do not check any byte
past the allocation length that was sent to the INQUIRY command.

Reported-by: Max Reitz <mreitz@redhat.com>
Reviewed-by: Max Reitz <mreitz@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 57dbb58d800f62b9e56d946660dba4e8dbd20204)
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>